backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiortc>=1.14.0",
    "fastapi>=0.121.3",
    "loguru>=0.7.3",
    "pipecat-ai[cartesia,deepgram,openai,runner,silero,webrtc]>=0.0.96",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.0.0",
    "uvicorn>=0.38.0",
]
